Move HttpUtil to .usagetracker
