+++ plugins/com.google.cloud.tools.eclipse.usagetracker.test/src/com/google/cloud/tools/eclipse/usagetracker/HttpUtilTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.eclipse.util.io;
+package com.google.cloud.tools.eclipse.usagetracker;
 
 import java.util.Collections;
 import java.util.HashMap;
