+++ plugins/com.google.cloud.tools.eclipse.usagetracker.test/src/com/google/cloud/tools/eclipse/usagetracker/HttpUtilWithServerTest.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.eclipse.util.io;
+package com.google.cloud.tools.eclipse.usagetracker;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import com.google.cloud.tools.eclipse.test.util.http.TestHttpServer;
+import com.google.cloud.tools.eclipse.usagetracker.HttpUtil;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
