+++ plugins/com.google.cloud.tools.eclipse.util/.classpath
@@ -4,6 +4,5 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src/"/>
-	<classpathentry kind="lib" path="lib/google-http-client-1.22.0.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
