+++ plugins/com.google.cloud.tools.eclipse.util/META-INF/MANIFEST.MF
@@ -13,9 +13,7 @@ Export-Package: com.google.cloud.tools.eclipse.util,
  com.google.cloud.tools.eclipse.util.templates.appengine
 Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
  com.google.common.base;version="[20.0.0,21.0.0)",
- com.google.common.escape;version="[20.0.0,21.0.0)",
  com.google.common.io;version="[20.0.0,21.0.0)",
- com.google.common.net;version="[20.0.0,21.0.0)",
  freemarker.template;version="[2.3.25,2.4.0)",
  org.eclipse.core.commands,
  org.eclipse.core.expressions,
@@ -27,7 +25,6 @@ Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
  org.eclipse.wst.common.project.facet.core,
  org.osgi.framework
 Bundle-ClassPath: .,
- templates/appengine/,
- lib/google-http-client-1.22.0.jar
+ templates/appengine/
 Require-Bundle: org.eclipse.m2e.maven.runtime,
  org.eclipse.m2e.core;bundle-version="1.6.2"
