+++ plugins/com.google.cloud.tools.eclipse.util/build.properties
@@ -3,7 +3,6 @@ output.. = target/classes/
 bin.includes = META-INF/,\
                .,\
                templates/appengine/,\
-               lib/google-http-client-1.22.0.jar,\
                plugin.xml
 javacSource=1.7
 javacTarget=1.7
